interdiction d'ajouter une commande existante + enlever le tag quand le bot répond

+++ apptest.js
@@ -6,6 +6,9 @@ const dotenv = require('dotenv');
 dotenv.config();
 
 const mySecret = process.env['token'];
+const admin = process.env['admin'];
+const myChannel1 = process.env['channel1'];
+const myChannel2 = process.env['channel2'];
 
 // Create a new client instance
 const client = new Client({ intents: [Intents.FLAGS.GUILDS, Intents.FLAGS.GUILD_MESSAGES] });
@@ -15,29 +18,29 @@ let myArr = []
 
 // add a new command 
 function addCommand(msg) {
-  //let re = /"(\w+)\W*(\w+)"/g
-  //myArr = msg.content.replace(re, '$1,$2').split(" ")
-  //console.log(msg.content.split(" ")[2])
   let strArr = ""
-  for (let i in msg.content.split(" ")) {
-    if (i < 2) {
-      myArr.push(msg.content.split(" ")[i])
-    } else {
-      strArr += " "+msg.content.split(" ")[i]
+  let msgExist = msg.content.split(" ")[1]
+  if (msgExist === "add" || msgExist === "help" || msgExist === "del" || msgExist === "pascontent" || msgExist === "list") {
+    msg.reply(`La commande **${msgExist}** existe déjà`)
+  } 
+  else { 
+    for (let i in msg.content.split(" ")) {
+      if (i < 2) {
+        myArr.push(msg.content.split(" ")[i])
+      } else {
+        strArr += " "+msg.content.split(" ")[i]
+      }
     }
-  }
-  myArr.push(strArr)
-  console.log(strArr)
-  console.log(myArr)
-  if (myArr.length != 3) {
-    msg.reply("max 2 args")
-  } else {
-    json[myArr[1]] = myArr[2].replace(",", " ");
-    fs.writeFile("./info.json", "\n"+JSON.stringify(json), err => {
-      if(err) console.log(err);
-    });
-    msg.reply(`Votre entrée **${myArr[1]}** a bien été ajoutée`)
-    myArr = []
+    if (myArr.length != 3) {
+      msg.reply("max 2 args")
+    } else {
+      json[myArr[1]] = myArr[2].replace(",", " ");
+      fs.writeFile("./info.json", "\n"+JSON.stringify(json), err => {
+        if(err) console.log(err);
+      });
+      msg.channel.send(`Votre entrée **${myArr[1]}** a bien été ajoutée`)
+      myArr = []
+    }
   }
 }
 
@@ -48,7 +51,7 @@ function listCommands(msg) {
     arrTest.push("$"+key)
   }
   let newArr = "```\n"+arrTest.toString().replace(/,/g, " \n")+"```\n"
-  msg.reply(`la liste : **${newArr}**`);
+  msg.channel.send(`**${newArr}**`);
 }
 
 // delete a command
@@ -62,7 +65,7 @@ function deleteCommand(msg) {
     fs.writeFile("./info.json", "\n"+JSON.stringify(json), err => {
       if(err) console.log(err);
     });
-    msg.reply(`Votre entrée **${myArr[1]}** a bien été supprimé`)
+    msg.channel.send(`La commande **${myArr[1]}** a bien été supprimé`)
   }
 }
 
@@ -83,12 +86,12 @@ client.once("ready", () => {
 
 client.on("messageCreate", msg => {
   // only the channel specified
-  if (msg.channel.id === "918606393219616798" || msg.channel.id === "918607399357669406") {
+  if (msg.channel.id === myChannel1 || msg.channel.id === myChannel2) {
     if (msg.author.bot) return;
     
     let userMessage = msg.content.slice(1);
     if(json.hasOwnProperty(userMessage)) {
-      msg.reply(json[userMessage])
+      msg.channel.send(json[userMessage])
       .then(() => console.log(`Replied to message ${msg.content}`))
       .catch(console.error);
     } 
@@ -102,12 +105,12 @@ client.on("messageCreate", msg => {
       deleteCommand(msg)
     }
     else if (msg.content === "$pascontent") {
-      if (msg.author.id === "202118392008802305") {
+      if (msg.author.id === admin) {
         clearChat(msg)
       }
     }
     else if (msg.content === "$help") {
-      msg.reply("```\nLE BOT N'EST ACTIF QUE SUR LE CHANNEL commands\n\nLa liste des commandes actuelles (évolue au fur & à mesure) :\n\n$list : Affiche toutes les commandes relatifs à la formation Devops, c'est aussi basé sur vos ajouts au fur et à mesure.\n\n$add : la syntaxe exacte c'est '$add macommande texte'\n- macommande = le nom que vous donnerez à la commande.\n- texte sera votre descriptif ou lien ou image ou autre.\nexemples :\n - $add share https://nc1.opencom.eu/index.php/s/oeoqFzgPT8aycLG?path=%2F12-Docker-Images\n - $add mdpshare DevOps21fai9Ee\n\n$del : la syntaxe '$del commande' : assez explicite ça supprime la commande présente dans la liste\n\n$help : Affiche ce message\n\n```")
+      msg.channel.send("```\nLE BOT N'EST ACTIF QUE SUR LE CHANNEL commands\n\nLa liste des commandes actuelles (évolue au fur & à mesure) :\n\n$list : Affiche toutes les commandes relatifs à la formation Devops, c'est aussi basé sur vos ajouts au fur et à mesure.\n\n$add : la syntaxe exacte c'est '$add macommande texte'\n- macommande = le nom que vous donnerez à la commande.\n- texte sera votre descriptif ou lien ou image ou autre.\nexemples :\n - $add share https://nc1.opencom.eu/index.php/s/oeoqFzgPT8aycLG?path=%2F12-Docker-Images\n - $add mdpshare DevOps21fai9Ee\n\n$del : la syntaxe '$del commande' : assez explicite ça supprime la commande présente dans la liste\n\n$help : Affiche ce message\n\n```")
     }
     else if (msg.content.slice(0,1) === "$") {
         msg.reply("La requête demandée n'existe pas, si vous désirez l'ajouter merci de taper $add")
